feat: Update MainNav component to include MobileMenu and MenuManager
This commit adds the MobileMenu and MenuManager components to the MainNav component. This allows for a mobile navigation menu to be displayed when the screen size is small. The DesktopNavigation component has been commented out for now. This change improves the user experience by providing a responsive navigation menu.

Refactor the code to include the MobileMenu and MenuManager components in the MainNav component.