refactor: Import DentallyPortal constant in DesktopNav component
This commit updates the DesktopNav component to import the `DentallyPortal` constant from the `@/lib/constants` module. This change ensures that the `DentallyPortal` value is used correctly in the `Link` component, allowing users to open the Dentally portal in a new tab when clicking the "Book Now" button.

Refactor the DesktopNav component to import `DentallyPortal` constant.