refactor: Update ServiceActions component to open Dentally portal in a new tab
This commit updates the ServiceActions component to open the Dentally portal in a new tab when the "Book Now" button is clicked. It imports the `DentallyPortal` constant from the `@/lib/constants` module and sets the `href` prop of the `Link` component to `${DentallyPortal}`. This change improves the user experience by providing a seamless way to access the Dentally portal without leaving the current page.

Refactor the ServiceActions component to open Dentally portal in a new tab.